refactor : dto 패키지 분리 및 Stream API 적용하여 수정
